# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Course Management Platform API",
    "version": "1.0.0",
    "description": "API documentation for the Course Management Platform backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/cohorts": {
      "get": {
        "summary": "Get all cohorts",
        "tags": [
          "Cohorts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of cohorts"
          }
        }
      },
      "post": {
        "summary": "Create a cohort",
        "tags": [
          "Cohorts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cohort"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cohort created"
          }
        }
      }
    },
    "/cohorts/{id}": {
      "get": {
        "summary": "Get a cohort by ID",
        "tags": [
          "Cohorts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Cohort object"
          }
        }
      },
      "put": {
        "summary": "Update a cohort",
        "tags": [
          "Cohorts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cohort"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cohort updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a cohort",
        "tags": [
          "Cohorts"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Cohort deleted"
          }
        }
      }
    },
    "/course-offerings": {
      "get": {
        "summary": "Get all course offerings",
        "tags": [
          "Course Offerings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of course offerings"
          }
        }
      },
      "post": {
        "summary": "Create a course offering",
        "tags": [
          "Course Offerings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseOffering"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Course offering created"
          }
        }
      }
    },
    "/course-offerings/{id}": {
      "get": {
        "summary": "Get a course offering by ID",
        "tags": [
          "Course Offerings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course offering object"
          }
        }
      },
      "put": {
        "summary": "Update a course offering",
        "tags": [
          "Course Offerings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseOffering"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course offering updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a course offering",
        "tags": [
          "Course Offerings"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course offering deleted"
          }
        }
      }
    },
    "/facilitators": {
      "get": {
        "summary": "Get all facilitators",
        "tags": [
          "Facilitators"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of facilitators"
          }
        }
      },
      "post": {
        "summary": "Create a facilitator",
        "tags": [
          "Facilitators"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Facilitator"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Facilitator created"
          }
        }
      }
    },
    "/facilitators/{id}": {
      "get": {
        "summary": "Get a facilitator by ID",
        "tags": [
          "Facilitators"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Facilitator object"
          }
        }
      },
      "put": {
        "summary": "Update a facilitator",
        "tags": [
          "Facilitators"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Facilitator"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Facilitator updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a facilitator",
        "tags": [
          "Facilitators"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Facilitator deleted"
          }
        }
      }
    },
    "/managers": {
      "get": {
        "summary": "Get all managers",
        "tags": [
          "Managers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of managers"
          }
        }
      },
      "post": {
        "summary": "Create a new manager",
        "tags": [
          "Managers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Manager created"
          }
        }
      }
    },
    "/managers/{id}": {
      "get": {
        "summary": "Get manager by ID",
        "tags": [
          "Managers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A manager object"
          }
        }
      },
      "put": {
        "summary": "Update a manager",
        "tags": [
          "Managers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Manager updated"
          }
        }
      },
      "delete": {
        "summary": "Delete a manager",
        "tags": [
          "Managers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Manager deleted"
          }
        }
      }
    },
...
  },
...
}